reorganize into subdirs
