+++ modules/create_extra/CMakeLists.txt
@@ -1,18 +1,15 @@
 if (${build_create_extra})
     message("building create extra")
 
-    # requirements
-    find_library(kipr_libs kipr REQUIRED)
-
     # sources
     set(SOURCES ${CMAKE_CURRENT_SOURCE_DIR}/src/create_extra.cpp)
-    include_directories(${CMAKE_CURRENT_SOURCE_DIR}/include)
+    set(EXTERNAL_DEPS kipr)
 
-    #library
-    add_library(ihs_create_extra OBJECT ${SOURCES})
-    target_link_libraries(ihs_create_extra PUBLIC ${kipr_libs})
+    add_module(create_extra "${SOURCES}")
+    add_external_deps(create_extra "${EXTERNAL_DEPS}")
 
     # install
     get_property(local_install_path GLOBAL PROPERTY install_path)
-    install(FILES ${CMAKE_CURRENT_SOURCE_DIR}/include/create_extra.hpp DESTINATION ${local_install_path})
+    set(install_file_dir ${CMAKE_CURRENT_SOURCE_DIR}/include/ihsboost/create_extra)
+    install(FILES ${install_file_dir}/create_extra.hpp DESTINATION ${local_install_path}/create_extra)
 endif()
