+++ modules/movement/CMakeLists.txt
@@ -1,9 +1,6 @@
 if (${build_movement})
     message("building movement")
 
-    # requirements
-    find_library(kipr_libs kipr REQUIRED)
-
     # sources
     set(SOURCES ${CMAKE_CURRENT_SOURCE_DIR}/src/accelerate.cpp 
                 ${CMAKE_CURRENT_SOURCE_DIR}/src/imu_movement.cpp 
@@ -12,29 +9,23 @@ if (${build_movement})
                 ${CMAKE_CURRENT_SOURCE_DIR}/src/roomba_movement.cpp
                 ${CMAKE_CURRENT_SOURCE_DIR}/src/singletons.cpp
                 )
-    include_directories(${CMAKE_CURRENT_SOURCE_DIR}/include)
+    set(DEPS util controllers threading)
+    set(EXTERNAL_DEPS kipr)
 
-    # include dependencies
-    include_directories(${CMAKE_SOURCE_DIR}/modules/util/include)
-    include_directories(${CMAKE_SOURCE_DIR}/modules/controllers/include)
-    include_directories(${CMAKE_SOURCE_DIR}/modules/threading/include)
-
-    # library
-    add_library(ihs_movement OBJECT ${SOURCES})
-    target_link_libraries(ihs_movement PUBLIC ${kipr_libs})
-
-    # library dependencies
-    add_dependencies(ihs_movement ihs_threading ihs_util ihs_controllers)
+    add_module(movement "${SOURCES}")
+    add_ihsboost_deps(movement "${DEPS}")
+    add_external_deps(movement "${EXTERNAL_DEPS}")
 
     # install
     get_property(local_install_path GLOBAL PROPERTY install_path)
-    install(FILES ${CMAKE_CURRENT_SOURCE_DIR}/include/accelerate.hpp 
-        ${CMAKE_CURRENT_SOURCE_DIR}/include/imu_movement.hpp 
-        ${CMAKE_CURRENT_SOURCE_DIR}/include/line_follow.hpp
-        ${CMAKE_CURRENT_SOURCE_DIR}/include/maneuvers.hpp 
-        ${CMAKE_CURRENT_SOURCE_DIR}/include/movement_constants.hpp
-        ${CMAKE_CURRENT_SOURCE_DIR}/include/movement.hpp
-        ${CMAKE_CURRENT_SOURCE_DIR}/include/roomba_movement.hpp
-        ${CMAKE_CURRENT_SOURCE_DIR}/include/singletons.hpp
-        DESTINATION ${local_install_path})
+    set(install_file_dir ${CMAKE_CURRENT_SOURCE_DIR}/include/ihsboost/movement/)
+    install(FILES ${install_file_dir}/accelerate.hpp 
+        ${install_file_dir}/imu_movement.hpp 
+        ${install_file_dir}/line_follow.hpp
+        ${install_file_dir}/maneuvers.hpp 
+        ${install_file_dir}/movement_constants.hpp
+        ${install_file_dir}/movement.hpp
+        ${install_file_dir}/roomba_movement.hpp
+        ${install_file_dir}/singletons.hpp
+        DESTINATION ${local_install_path}/movement)
 endif()
