+++ modules/servos/CMakeLists.txt
@@ -6,20 +6,15 @@ if (${build_servos})
 
     # sources
     set(SOURCES ${CMAKE_CURRENT_SOURCE_DIR}/src/servos.cpp)
-    include_directories(${CMAKE_CURRENT_SOURCE_DIR}/include)
+    set(DEPS controllers util)
+    set(EXTERNAL_DEPS kipr)
 
-    # include dependencies
-    include_directories(${CMAKE_SOURCE_DIR}/modules/controllers/include)
-    include_directories(${CMAKE_SOURCE_DIR}/modules/util/include)
-
-    # library
-    add_library(ihs_servos OBJECT ${SOURCES})
-    target_link_libraries(ihs_servos PUBLIC ${kipr_libs})
-
-    # library dependencies
-    add_dependencies(ihs_servos ihs_controllers)
+    add_module(servos "${SOURCES}")
+    add_ihsboost_deps(servos "${DEPS}")
+    add_external_deps(servos "${EXTERNAL_DEPS}")
 
     # install
     get_property(local_install_path GLOBAL PROPERTY install_path)
-    install(FILES ${CMAKE_CURRENT_SOURCE_DIR}/include/servos.hpp DESTINATION ${local_install_path})
+    set(install_file_dir ${CMAKE_CURRENT_SOURCE_DIR}/include/ihsboost/servos)
+    install(FILES ${install_file_dir}/servos.hpp DESTINATION ${local_install_path}/servos)
 endif()
