+++ modules/util/CMakeLists.txt
@@ -1,24 +1,21 @@
 if (${build_util})
     message("building util")
 
-    # requirements
-    find_library(json_libs jsoncpp REQUIRED)
-
     # sources
     set(SOURCES ${CMAKE_CURRENT_SOURCE_DIR}/src/speed.cpp ${CMAKE_CURRENT_SOURCE_DIR}/src/timer.cpp
         ${CMAKE_CURRENT_SOURCE_DIR}/src/json_config.cpp ${CMAKE_CURRENT_SOURCE_DIR}/src/config.cpp)
-    include_directories(${CMAKE_CURRENT_SOURCE_DIR}/include)
+    set(EXTERNAL_DEPS jsoncpp)
 
-    # library
-    add_library(ihs_util OBJECT ${SOURCES})
-    target_link_libraries(ihs_util PUBLIC ${json_libs})
+    add_module(util "${SOURCES}")
+    add_external_deps(util "${EXTERNAL_DEPS}")
 
     # install
     get_property(local_install_path GLOBAL PROPERTY install_path)
-    install(FILES ${CMAKE_CURRENT_SOURCE_DIR}/include/cliff.hpp ${CMAKE_CURRENT_SOURCE_DIR}/include/speed.hpp 
-        ${CMAKE_CURRENT_SOURCE_DIR}/include/timer.hpp ${CMAKE_CURRENT_SOURCE_DIR}/include/roomba_config.hpp 
-        ${CMAKE_CURRENT_SOURCE_DIR}/include/small_bot_config.hpp ${CMAKE_CURRENT_SOURCE_DIR}/include/config.hpp 
-        ${CMAKE_CURRENT_SOURCE_DIR}/include/json_config.hpp ${CMAKE_CURRENT_SOURCE_DIR}/include/roomba_config.json 
-        ${CMAKE_CURRENT_SOURCE_DIR}/include/small_bot_config.json ${CMAKE_CURRENT_SOURCE_DIR}/include/util.hpp
-        DESTINATION ${local_install_path})
+    set(install_file_dir ${CMAKE_CURRENT_SOURCE_DIR}/include/ihsboost/util)
+    install(FILES ${install_file_dir}/cliff.hpp ${install_file_dir}/speed.hpp 
+        ${install_file_dir}/timer.hpp ${install_file_dir}/roomba_config.hpp 
+        ${install_file_dir}/small_bot_config.hpp ${install_file_dir}/config.hpp 
+        ${install_file_dir}/json_config.hpp ${install_file_dir}/roomba_config.json 
+        ${install_file_dir}/small_bot_config.json ${install_file_dir}/util.hpp
+        DESTINATION ${local_install_path}/util)
 endif()
