+++ modules/util/include/ihsboost/util/small_bot_config.json
